--- src/components/Paintings.js
@@ -1,56 +1,0 @@
-import React from 'react';
-import Overlay from './Overlay';
-import '../stylesheet/Gallery.css';
-var slugify = require('slugify')
-
-const Paintings = (props) =>  {
-  const data = props.paintingsData;
-      return(
-        <div className="gallery">
-        {props.closed===true
-        ?
-        ""
-        :
-        <Overlay info={props.info} galleryType={props.galleryType} handleArrowClick={props.handleArrowClick} handleCloseClick={props.handleCloseClick}/>
-        }
-          <div className="row-grid">
-            <div className="column-grid">
-            {
-           [  '25', '26', '1', '19', '0', '2', '21'].map((i, count)=>{
-            return <div key={i} className=""><img src={require(`../imgs/paintings/${data[i].url}`)} id={slugify(data[i].name)} title={data[i].name} style={{width:'100%'}} name="paintings" alt={data[i].name} onClick={props.handleImageClick}/><div className="mobile-info">{data[i].sold ? <p className="sold-mobile" onClick={props.handleCloseClick}>SOLD</p> : <a href='/contact'><p className="label-mobile" onClick={props.handleCloseClick}>BUY</p></a>}<p className="all-caps">{data[i].name}</p><i><p className="title">{data[i].specs}</p></i><p className="desc">{data[i].description}</p><hr /></div></div>
-            
-          })
-          }
-         </div>
-            <div className="column-grid">
-            {
-           [  '22','23', '20', '18', '15', '4'].map((i, count)=>{
-             return <div key={i}><img src={require(`../imgs/paintings/${data[i].url}`)} id={slugify(data[i].name)} title={data[i].name} style={{width:'100%'}} name="paintings" alt={data[i].name} onClick={props.handleImageClick}/><div className="mobile-info">{data[i].sold ? <p className="sold-mobile" onClick={props.handleCloseClick}>SOLD</p> : <a href='/contact'><p className="label-mobile" onClick={props.handleCloseClick}>BUY</p></a>}<p className="all-caps">{data[i].name}</p><i><p className="title">{data[i].specs}</p></i><p className="desc">{data[i].description}</p><hr /></div></div>
-           })
-          }
-           
-            </div>
-            <div className="column-grid">
-            {
-           ['8', '24', '5', '6', '16', '14', '7'].map((i, count)=>{
-             return <div key={i}><img src={require(`../imgs/paintings/${data[i].url}`)} title={data[i].name} id={slugify(data[i].name)} style={{width:'100%'}} name="paintings" alt={data[i].name} onClick={props.handleImageClick}/><div className="mobile-info">{data[i].sold ? <p className="sold-mobile" onClick={props.handleCloseClick}>SOLD</p> : <a href='/contact'><p className="label-mobile" onClick={props.handleCloseClick}>BUY</p></a>}<p className="all-caps">{data[i].name}</p><i><p className="title">{data[i].specs}</p></i><p className="desc">{data[i].description}</p><hr /></div></div>
-             
-           })
-          }
-
-            </div>
-            <div className="column-grid">
-            {
-           ['10', '17','11', '12', '13', '9', '3'].map((i, count)=>{
-             return <div key={i}><img src={require(`../imgs/paintings/${data[i].url}`)} title={data[i].name} id={slugify(data[i].name)} style={{width:'100%'}} name="paintings" alt={data[i].name} onClick={props.handleImageClick}/><div className="mobile-info">{data[i].sold ? <p className="sold-mobile" onClick={props.handleCloseClick}>SOLD</p> : <a href='/contact'><p className="label-mobile" onClick={props.handleCloseClick}>BUY</p></a>}<p className="all-caps">{data[i].name}</p><i><p className="title">{data[i].specs}</p></i><p className="desc">{data[i].description}</p><hr /></div></div>
-             
-           })
-          }
-              
-            </div>
-          </div>
-        </div>
-      );
-};
-
-export default Paintings;
