+++ src/data/drawings.json
@@ -53,7 +53,7 @@
     },
     {
       "type": "drawings",
-      "name": "Tree",
+      "name": "Tree 2",
       "description": "",
       "url": "Tree2.jpg",
       "specs": ""
