+++ src/data/illustrations.json
@@ -148,7 +148,7 @@
     },
     {
       "type": "illustrations",
-      "name": "turtle",
+      "name": "turtle 2",
       "description": "",
       "url": "turtle2.jpg",
       "specs": ""
@@ -239,7 +239,7 @@
     },
     {
       "type": "illustrations",
-      "name": "Fish",
+      "name": "Fish 2",
       "description": "",
       "url": "fish2.jpg",
       "specs": ""
@@ -323,7 +323,7 @@
     },
     {
       "type": "illustrations",
-      "name": "Fish",
+      "name": "Fishy",
       "description": "",
       "url": "fish_olenka.jpg",
       "specs": ""
@@ -400,7 +400,7 @@
     },
     {
       "type": "illustrations",
-      "name": "Butterfly",
+      "name": "Butterfly 2",
       "description": "",
       "url": "butterfly.jpg",
       "specs": ""
