--- src/stylesheet/Gallery.css
@@ -1,95 +1,0 @@
-@media (min-width: 480px) {
-  .mobile-info{
-    display: none;
-  }
-}
-@media (max-width: 480px) {
-  .sold-mobile {
-    position: absolute;
-    height: 32px;
-    color: white;
-    background-color: rgba(236, 42, 42, 0.65);
-    border-radius: 10px;
-    padding: 8px 10px;
-    right: 15px;
-    margin: 0;
-    font-size: 12px;
-    margin-top: -5px;
-  }
-  
-  .label-mobile {
-    position: absolute;
-    height: 32px;
-    color: white;
-    background-color: rgba(53, 197, 17, 0.75);
-    border-radius: 10px;
-    padding: 9px 16px;
-    right: 15px;
-    margin: 0;
-    font-size: 12px;
-    margin-top: -5px;
-  }
-  .title {
-    line-height: 0px;
-  }
-  .all-caps{
-    font-weight: bold;
-  }
-  p.desc {
-    margin-top: 10px !important;
-    margin-bottom: 15px !important;
-  }
-
-  hr{
-    width: 10%;
-
-  }
-}
-
-.gallery *{
-  box-sizing: border-box;
-}
-
-.row-grid {
-  display: -ms-flexbox; /* IE10 */
-  display: flex;
-  -ms-flex-wrap: wrap; /* IE10 */
-  flex-wrap: wrap;
-  padding: 0 4px;
-  width: 100%;
-}
-
-/* Create four equal columns that sits next to each other */
-.column-grid {
-  -ms-flex: 25%; /* IE10 */
-  flex: 25%;
-  max-width: 25%;
-  padding: 0 4px;
-}
-
-.column-grid img {
-  margin-top: 8px;
-  vertical-align: middle;
-}
-
-/* Responsive layout - makes a two column-layout instead of four columns */
-@media screen and (max-width: 1200px) {
-  .column-grid {
-    -ms-flex: 50%;
-    flex: 50%;
-    max-width: 50%;
-  }
-}
-
-/* Responsive layout - makes the two columns stack on top of each other instead of next to each other */
-@media screen and (max-width: 600px) {
-  .column-grid {
-    -ms-flex: 100%;
-    flex: 100%;
-    max-width: 100%;
-  }
-}
-
-.gallery p{
-  text-align: center;
-}
