+++ src/stylesheet/Gallery.scss
